Use os-release(5) to determine the OS name.
Create a global `os_release()` function that returns a dict containing the
contents of the os-release file (use shlex to parse it).  Accompanying this
is the `OS_RELEASE` global, since I figure calling it multiple times is a
waste.  I'm not really sure that this was a good place to put it, but I
didn't have a good sense of how you were trying to organize the program.
Perhaps pass it around in the `state`?

Now, because the old code in distroCheck.render was basically hard-coding
the string "Arch Linux", but pretending to actually detect something, the
code that used it to select the logo was a bit fragile.  Instead, create
a .logo() method on the Archey object that finds the most appropriate logo
based on some logic describe below.  This is better than trying to parse
the output of the distro module.

Archey.logo() goes through the PRETTY_NAME, NAME, ID, `uname -o`, and
finally an empty string, until it finds a logo matching that string.  This
way we can be pretty robust in finding the most appropriate logo.  I
included the empty string at the end as an absolute; fallback and added a
logo that is a giant "?" for it.  This will allow, say, a "GNU/Linux" logo
to be added that will show on otherwise unknown GNU/Linux distros, while
still only showing the "?" logo on something else, like BSD.

Finally, the distroCheck.render() code was updated to try PRETTY_NAME,
NAME, then `uname -o`.