Better fraction detection.
Previously, it looked for exactly one slash in a value to identify a
fraction.  This resulted in silly things like "GNU/Linux" being rendered as
a fraction (the "GNU" becomes blue, or whatever the color is set to).

Now, it inspects that there is exactly one slash, with at least one digit
on each side of it.  I didn't mandate entirely digits, because a fraction
might include units.