Make ram() option locale independent
I updated the ram module to make it independent from the users locale.
Before it used the free(1) command and looked for the String "Mem:" to
detect the desired output values, but this aproach will break on
any non-english system.
Now the module reads the data from /proc/meminfo and calculates
the desired values on its own, just like free(1) does.

Possible improvements:
* Let python read the file directly instead of cat'ing it.